fix: studio endpoint to support lazy project loading
