+++ studio/pages/api/projects/index.ts
@@ -26,39 +26,7 @@ const handleGetAll = async (req: NextApiRequest, res: NextApiResponse) => {
       cloud_provider: 'localhost',
       status: 'ACTIVE_HEALTHY',
       region: 'local',
-      connectionString: createDbConnectionString({
-        db_user_supabase: 'postgres',
-        db_host: 'localhost',
-        db_pass_supabase: process.env.POSTGRES_PASSWORD as string,
-        db_port: 5432,
-        db_name: 'postgres',
-        db_ssl: false,
-      }),
     },
   ]
   return res.status(200).json(response)
-}
-
-/**
- * Creates a Postgres connection string using the Supabase master login.
- * Expects the passwords to be encrypted (straight from the DB)
- */
-const createDbConnectionString = ({
-  db_user_supabase,
-  db_pass_supabase,
-  db_host,
-  db_port,
-  db_name,
-  db_ssl,
-}: {
-  db_user_supabase: string
-  db_host: string
-  db_pass_supabase: string
-  db_port: number
-  db_name: string
-  db_ssl: boolean
-}) => {
-  return `postgres://${db_user_supabase}:${db_pass_supabase}@${db_host}:${db_port}/${db_name}?sslmode=${
-    db_ssl ? 'require' : 'disable'
-  }`
 }
